added MessageListContext for prop management
