+++ likeminds-chat-reactnative-integration/ChatSX/screens/ChatRoom/index.tsx
@@ -561,102 +561,7 @@ const ChatroomComponent = () => {
 
   return (
     <View style={styles.container}>
-      {shimmerIsLoading ? (
-        <View style={{ marginTop: 10 }}>
-          <View
-            style={{
-              backgroundColor: "#e8e8e877",
-              width: 200,
-              paddingLeft: 8,
-              paddingVertical: 15,
-              borderTopRightRadius: 12,
-              borderTopLeftRadius: 12,
-              borderBottomRightRadius: 12,
-            }}
-          >
-            <ShimmerPlaceHolder
-              style={{ width: 150, height: 10, borderRadius: 5 }}
-            />
-            <ShimmerPlaceHolder
-              style={{ width: 120, height: 10, marginTop: 10, borderRadius: 5 }}
-            />
-          </View>
-
-          <View
-            style={{
-              backgroundColor: "#e8e8e877",
-              alignSelf: "flex-end",
-              width: 200,
-              paddingLeft: 8,
-              paddingVertical: 15,
-              borderTopRightRadius: 12,
-              borderTopLeftRadius: 12,
-              borderBottomLeftRadius: 12,
-              marginTop: 10,
-            }}
-          >
-            <ShimmerPlaceHolder
-              style={{ width: 150, height: 10, borderRadius: 5 }}
-            />
-            <ShimmerPlaceHolder
-              style={{ width: 120, height: 10, marginTop: 10, borderRadius: 5 }}
-            />
-          </View>
-          <View
-            style={{
-              backgroundColor: "#e8e8e877",
-              width: 200,
-              paddingLeft: 8,
-              paddingVertical: 15,
-              borderTopRightRadius: 12,
-              borderTopLeftRadius: 12,
-              borderBottomRightRadius: 12,
-              marginTop: 10,
-            }}
-          >
-            <ShimmerPlaceHolder
-              style={{ width: 150, height: 10, borderRadius: 5 }}
-            />
-            <ShimmerPlaceHolder
-              style={{ width: 120, height: 10, marginTop: 10, borderRadius: 5 }}
-            />
-          </View>
-
-          <View
-            style={{
-              backgroundColor: "#e8e8e877",
-              alignSelf: "flex-end",
-              width: 200,
-              paddingLeft: 8,
-              paddingVertical: 15,
-              borderTopRightRadius: 12,
-              borderTopLeftRadius: 12,
-              borderBottomLeftRadius: 12,
-              marginTop: 10,
-            }}
-          >
-            <ShimmerPlaceHolder
-              style={{ width: 150, height: 10, borderRadius: 5 }}
-            />
-            <ShimmerPlaceHolder
-              style={{ width: 120, height: 10, marginTop: 10, borderRadius: 5 }}
-            />
-          </View>
-        </View>
-      ) : (
-        <>
-          <MessageList
-            chatroomID={chatroomID}
-            handleLongPress={handleLongPress}
-            handleClick={handleClick}
-            removeReaction={removeReaction}
-            onTapToUndo={onTapToUndo}
-            handleFileUpload={handleFileUpload}
-            navigation={navigation}
-            ref={refInput}
-          />
-        </>
-      )}
+      <MessageList />
 
       {/* Input Box Flow */}
       <View
